+++ eslint.config.cjs
@@ -1,13 +1,8 @@
-export default [
+module.exports = [
   {
+    ...require('eslint-config-love'),
     files: ['src/**.ts'],
-    extends: [
-      'plugin:@typescript-eslint/recommended',
-      'plugin:@typescript-eslint/recommended-requiring-type-checking',
-      'standard'
-    ],
-    parser: '@typescript-eslint/parser',
-    plugins: ['@typescript-eslint'],
+    ignores: ['dist/**', 'node_modules/**'],
     rules: {
       '@typescript-eslint/member-delimiter-style': ['error', {
         multiline: {
@@ -19,6 +14,7 @@ export default [
           requireLast: false
         }
       }],
+      '@typescript-eslint/method-signature-style': 'off',
       '@typescript-eslint/no-explicit-any': 'off',
       '@typescript-eslint/no-non-null-assertion': 'off',
       '@typescript-eslint/no-unused-vars': ['error', { argsIgnorePattern: '^_' }],
@@ -27,13 +23,6 @@ export default [
         alphabetize: { order: 'asc', caseInsensitive: true },
         groups: ['index', 'sibling', 'parent', 'internal', 'external', 'builtin', 'object', 'type']
       }]
-    },
-    parserOptions: {
-      project: ['./tsconfig.json']
-    },
-    ignores: [
-      'dist/**',
-      'node_modules/**'
-    ]
+    }
   }
 ]